chore: move mandatory and shared generator code to common.ts
